fix(ui): one hostHwFlags for both drawers, not two that disagree
The profile drawer carried its own private copy of the host-capability
translator, written before the kfd-aware ROCm fix (#2188). On a box with
/dev/kfd but no host rocminfo — the exact fleet case #2188 fixed — the slot
drawer offered ROCm runtimes and the profile drawer hid them. It also
coerced a hardware payload with no gpus[0] into an explicit all-false veto,
where the slot drawer reads that shape as "unknown" and vetoes nothing.

Move the helper to hw-cascade.js, next to the LANE_HW table and the
runnerOptions() filter that consume its output, and import it from both
drawers. hw-cascade.js is the module both already import and is cycle-free:
slot-modals.jsx imports runtimeChips FROM profiles.jsx, so the profile
drawer could never have imported the helper from the slot drawer.

Tests: two profile-drawer parity cases (kfd-present box keeps ROCm-only
runtimes; gpus[]-less payload vetoes nothing), both failing before the move.
