+++ ui/src/dash/__tests__/host-hw-flags.test.ts
@@ -1,4 +1,4 @@
-// hostHwFlags (slot-modals.jsx) — unknown-never-vetoes regression coverage.
+// hostHwFlags (hw-cascade.js) — unknown-never-vetoes regression coverage.
 //
 // hw-cascade.js's runnerOptions() hides a runtime only when a lane's flag
 // reads explicitly `false`; it never hides one on an absent/undefined flag
@@ -8,18 +8,13 @@
 // about. Caught by the Task 12 e2e mocks (a bare `hardware: {}` — no
 // gpus[0] — silently vetoed every GPU runtime); see task-12-report.md.
 //
-// slot-modals.jsx is a window-globals dash module (`const {...} = React` at
-// module top, `Object.assign(window, {...})` at the bottom), so the globals
-// must be installed before the dynamic import — same pattern as
-// runner-images-view.test.tsx / memoryOverviewV2.smoke.test.tsx.
-import React from 'react'
+// It lives in hw-cascade.js (a pure-functions module, no window globals)
+// rather than in slot-modals.jsx because both drawers consume it: the slot
+// drawer AND the profile drawer's Runtime select. The profile-drawer end of
+// that parity is asserted in profiles.test.tsx.
 import { describe, expect, it } from 'vitest'
 
-;(globalThis as unknown as { window: typeof globalThis }).window = globalThis
-;(globalThis as unknown as { React: typeof React }).React = React
-;(globalThis as unknown as { Icons: unknown }).Icons = new Proxy({}, { get: () => null })
-
-const { hostHwFlags } = await import('../slot-modals.jsx')
+import { hostHwFlags } from '../hw-cascade.js'
 
 describe('hostHwFlags', () => {
   it('no hardware payload at all (still loading) → {} (unknown, no veto)', () => {
